:suspect: extract ListBoxHeader to separate file
