+++ packages/layout/listbox/src/index.ts
@@ -1,3 +1,4 @@
 export * from './listbox'
+export * from './listbox-header'
 export * from './listbox-item'
 export * from './listbox-section'
