+++ packages/layout/listbox/src/listbox-section.tsx
@@ -2,14 +2,9 @@
 
 import type { StyleXStyles } from '@stylexjs/stylex'
 import * as stylex from '@stylexjs/stylex'
-import { Text } from '@urban-ui/text'
-import { tone } from '@urban-ui/theme/colors.stylex'
-import { control, space } from '@urban-ui/theme/layout.stylex'
+import { space } from '@urban-ui/theme/layout.stylex'
 import type { ListBoxSectionProps as AriaListBoxSectionProps } from 'react-aria-components'
-import {
-  Header as AriaHeader,
-  ListBoxSection as AriaListBoxSection,
-} from 'react-aria-components'
+import { ListBoxSection as AriaListBoxSection } from 'react-aria-components'
 
 import { useListBoxContext } from './listbox-context'
 
@@ -53,7 +48,7 @@ export interface ListBoxSectionProps<T extends object>
  * @example
  * ```tsx
  * <ListBoxSection>
- *   <Header>Fruits</Header>
+ *   <ListBoxHeader>Fruits</ListBoxHeader>
  *   <ListBoxItem id="apple">Apple</ListBoxItem>
  *   <ListBoxItem id="banana">Banana</ListBoxItem>
  * </ListBoxSection>
@@ -78,77 +73,3 @@ export function ListBoxSection<T extends object>({
   )
 }
 ListBoxSection.displayName = '@urban-ui/listbox-section'
-
-// Header styles
-const headerStyles = stylex.create({
-  header: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    paddingInline: space['100'],
-    color: tone.fgLo,
-  },
-})
-
-const headerSizeStyles = stylex.create({
-  md: {
-    paddingBlock: space['100'],
-    minHeight: control.md,
-  },
-  lg: {
-    paddingBlock: space['150'],
-    minHeight: control.lg,
-  },
-})
-
-export interface HeaderProps
-  extends Omit<React.HTMLAttributes<HTMLElement>, 'style' | 'className'> {
-  /**
-   * Size variant affects padding and height
-   * @default 'md'
-   */
-  size?: 'md' | 'lg'
-
-  /**
-   * Additional styles to apply
-   */
-  style?: StyleXStyles
-}
-
-/**
- * Visual label for a ListBoxSection. Sets correct padding for alignment with items.
- *
- * @example
- * ```tsx
- * <ListBoxSection>
- *   <ListBoxHeader>Category Name</ListBoxHeader>
- *   <ListBoxItem>...</ListBoxItem>
- * </ListBoxSection>
- * ```
- */
-export function ListBoxHeader({
-  size: sizeProp,
-  style,
-  children,
-  ...props
-}: HeaderProps) {
-  const context = useListBoxContext()
-  const size = sizeProp ?? context?.size ?? 'md'
-  const textSize = size === 'md' ? 'xs' : 'sm'
-
-  return (
-    <AriaHeader
-      {...props}
-      {...stylex.props(headerStyles.header, headerSizeStyles[size], style)}
-    >
-      {typeof children === 'string' ? (
-        <Text size={textSize} weight="semibold" tracking="wider" color="lo">
-          {children}
-        </Text>
-      ) : (
-        children
-      )}
-    </AriaHeader>
-  )
-}
-ListBoxHeader.displayName = '@urban-ui/listbox-header'
